Renamed UserConfig to Config
As UserConfig was always a bit of a misnomer - it is in fact
all the config that turns the collection of completely generic modules
into the desired solution, not just the bits that the end user
may take interest in. If motivation arises this could be split
into SolutionConfig and UserConfig,
but for now I prefer not to hide layers of config from the user.